Removing oclif from gatekeeper lib to make dependency list smaller
