+++ solana/gatekeeper-cli/.eslintignore
@@ -1,0 +1,2 @@
+dist/
+docs/
