+++ solana/gatekeeper-cli/.eslintrc
@@ -5,10 +5,15 @@
   "plugins": ["@typescript-eslint"],
   "extends": [
     "plugin:@typescript-eslint/recommended-requiring-type-checking",
+    "oclif",
+    "oclif-typescript",
     "prettier"
   ],
   "rules": {
-    "@typescript-eslint/no-implicit-any-catch": "error"
+    "@typescript-eslint/no-implicit-any-catch": "error",
+    "node/no-missing-import": ["error", {
+      "tryExtensions": [".js", ".ts", ".node"]
+    }]
   }
 }
