+++ solana/gatekeeper-cli/bin/dev.cmd
