+++ solana/gatekeeper-cli/bin/run.cmd
