+++ solana/gatekeeper-cli/docs/assets/search.js
@@ -1,0 +1,1 @@
+window.searchData = JSON.parse("{\"kinds\":{},\"rows\":[],\"index\":{\"version\":\"2.3.9\",\"fields\":[\"name\",\"comment\"],\"fieldVectors\":[],\"invertedIndex\":[],\"pipeline\":[]}}");
