+++ solana/gatekeeper-cli/package.json
@@ -1,0 +1,70 @@
+{
+  "name": "@identity.com/solana-gatekeeper-cli",
+  "description": "CLI to manage Gateway Tokens",
+  "version": "0.0.1",
+  "main": "dist/index.js",
+  "author": "dankelleher @dankelleher",
+  "license": "MIT",
+  "bin": {
+    "gateway": "./bin/run"
+  },
+  "dependencies": {
+    "@identity.com/solana-gatekeeper-lib": "4.1.1-alpha.3",
+    "@identity.com/solana-gateway-ts": "0.11.0-alpha.3",
+    "@oclif/core": "^1.20.4",
+    "@oclif/dev-cli": "^1.26.10",
+    "@oclif/plugin-help": "^5.1.19",
+    "@oclif/plugin-plugins": "^2.1.7",
+    "@oclif/plugin-version": "^1.1.3",
+    "@oclif/test": "^2.2.13",
+    "@solana/web3.js": "^1.69.0",
+    "eslint-config-oclif": "^4.0.0",
+    "eslint-config-oclif-typescript": "^1.0.3",
+    "oclif": "^3.4.0"
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^5.45.0",
+    "@typescript-eslint/parser": "^5.45.0",
+    "eslint": "^7.26.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "prettier": "^2.8.0",
+    "ts-node": "^10.9.1",
+    "typedoc": "^0.23.21",
+    "typescript": "^4.9.3"
+  },
+  "files": [
+    "/bin",
+    "/dist",
+    "/npm-shrinkwrap.json",
+    "/oclif.manifest.json"
+  ],
+  "oclif": {
+    "commands": "./dist/commands",
+    "additionalHelpFlags": [
+      "-h"
+    ],
+    "additionalVersionFlags": [
+      "-v"
+    ],
+    "bin": "gateway",
+    "topicSeparator": " ",
+    "plugins": [
+      "@oclif/plugin-help",
+      "@oclif/plugin-version",
+      "@oclif/plugin-plugins"
+    ]
+  },
+  "scripts": {
+    "build": "rm -rf dist && tsc -b && oclif-dev manifest && oclif-dev readme && cp src/util/*.json dist/util/ && yarn doc",
+    "postpack": "rm -f oclif.manifest.json",
+    "posttest": "yarn lint",
+    "prepack": "yarn build",
+    "lint": "eslint '{src,test}/**/*.ts' --max-warnings 0",
+    "lint:fix": "yarn lint --fix",
+    "version": "oclif-dev readme && git add README.md",
+    "doc": "yarn typedoc src/index.ts --readme none --excludeNotDocumented",
+    "audit-ci": "audit-ci --config audit-ci.json"
+  },
+  "types": "dist/index.d.ts"
+}
