+++ solana/gatekeeper-cli/src/commands/add-gatekeeper.ts
@@ -1,20 +1,20 @@
+import { airdropTo, GatekeeperNetworkService } from "@identity.com/solana-gatekeeper-lib";
 import { Command, Flags } from "@oclif/core";
+import { Keypair, PublicKey } from "@solana/web3.js";
+
 import {
   airdropFlag,
   clusterFlag,
   gatekeeperKeyFlag,
   gatekeeperNetworkKeyFlag,
-} from "../util/oclif/flags";
-import { Keypair, PublicKey } from "@solana/web3.js";
-import { airdropTo } from "../util";
-import { GatekeeperNetworkService } from "../service";
-import { getConnectionFromEnv } from "../util/oclif/utils";
+} from "../util/flags";
+import { getConnectionFromEnv } from "../util/utils";
 
-export default class RevokeGatekeeper extends Command {
-  static description = "Revoke a gatekeeper from a network";
+export default class AddGatekeeper extends Command {
+  static description = "Add a gatekeeper to a network";
 
   static examples = [
-    `$ gateway revoke-gatekeeper tgky5YfBseCvqehzsycwCG6rh2udA4w14MxZMnZz9Hp
+    `$ gateway add-gatekeeper tgky5YfBseCvqehzsycwCG6rh2udA4w14MxZMnZz9Hp
 `,
   ];
 
@@ -30,20 +30,20 @@ export default class RevokeGatekeeper extends Command {
     {
       name: "address",
       required: true,
-      description: "The address of the gatekeeper to revoke from the network",
+      description: "The address of the gatekeeper to add to the network",
       // eslint-disable-next-line @typescript-eslint/require-await
       parse: async (input: string): Promise<PublicKey> => new PublicKey(input),
     },
   ];
 
   async run(): Promise<void> {
-    const { args, flags } = await this.parse(RevokeGatekeeper);
+    const { args, flags } = await this.parse(AddGatekeeper);
 
     const gatekeeper: PublicKey = args.address as PublicKey;
     const gatekeeperNetwork = flags.gatekeeperNetworkKey as Keypair;
-    this.log(`Revoking: 
-      gatekeeper ${gatekeeper.toBase58()}
-      from network ${gatekeeperNetwork.publicKey.toBase58()}`);
+    this.log(`Adding:
+      gatekeeper ${gatekeeper.toBase58()} 
+      to network ${gatekeeperNetwork.publicKey.toBase58()}`);
 
     const connection = getConnectionFromEnv(flags.cluster);
 
@@ -56,19 +56,19 @@ export default class RevokeGatekeeper extends Command {
     }
 
     const networkService = new GatekeeperNetworkService(
-        connection,
-        gatekeeperNetwork
+      connection,
+      gatekeeperNetwork
     );
     const gatekeeperAccount = await networkService
-        .revokeGatekeeper(gatekeeper)
-        .then((t) => t.send())
-        .then((t) => t.confirm());
+      .addGatekeeper(gatekeeper)
+      .then((t) => t.send())
+      .then((t) => t.confirm());
     this.log(
-        `Revoked gatekeeper from network. Gatekeeper account: ${
-            gatekeeperAccount
-                ? gatekeeperAccount.toBase58()
-                : "//Gatekeeper Account Undefined//"
-        }`
+      `Added gatekeeper to network. Gatekeeper account: ${
+        gatekeeperAccount
+          ? gatekeeperAccount?.toBase58()
+          : "//GatekeeperAccount was undefined//"
+      }`
     );
   }
 }
