+++ solana/gatekeeper-cli/src/commands/refresh.ts
@@ -5,15 +5,15 @@ import {
   clusterFlag,
   gatekeeperKeyFlag,
   gatekeeperNetworkPubkeyFlag,
-} from "../util/oclif/flags";
-import { getTokenUpdateProperties } from "../util/oclif/utils";
+} from "../util/flags";
+import { getTokenUpdateProperties } from "../util/utils";
 
-export default class Freeze extends Command {
-  static description = "Freeze a gateway token";
+export default class Refresh extends Command {
+  static description = "Refresh a gateway token";
 
   static examples = [
-    `$ gateway freeze EzZgkwaDrgycsiyGeCVRXXRcieE1fxhGMp829qwj5TMv
-Frozen
+    `$ gateway refresh EzZgkwaDrgycsiyGeCVRXXRcieE1fxhGMp829qwj5TMv 54000
+Refreshed
 `,
   ];
 
@@ -33,23 +33,34 @@ Frozen
       // eslint-disable-next-line @typescript-eslint/require-await
       parse: async (input: string): Promise<PublicKey> => new PublicKey(input),
     },
+    {
+      name: "expiry",
+      description:
+        "The new expiry time in seconds for the gateway token (default 15 minutes)",
+      default: 15 * 60 * 60, // 15 minutes
+      // eslint-disable-next-line @typescript-eslint/require-await
+      parse: async (input: string): Promise<number> => Number(input),
+    },
   ];
 
   async run(): Promise<void> {
-    const { args, flags } = await this.parse(Freeze);
+    const { args, flags } = await this.parse(Refresh);
 
     const { gatewayToken, gatekeeper, service } =
       await getTokenUpdateProperties(args, flags);
 
-    this.log(`Freezing:
+    this.log(`Refreshing:
      ${gatewayToken.toBase58()}
      by gatekeeper ${gatekeeper.publicKey.toBase58()}`);
 
-    const frozenToken = await service
-      .freeze(gatewayToken)
+    await service
+      .updateExpiry(
+        gatewayToken,
+        Number.parseInt(args.expiry, 10) + Math.floor(Date.now() / 1000)
+      )
       .then((t) => t.send())
       .then((t) => t.confirm());
 
-    this.log("Frozen token", frozenToken?.publicKey.toBase58());
+    this.log("Refreshed");
   }
 }
