+++ solana/gatekeeper-cli/src/index.ts
@@ -1,0 +1,2 @@
+export { run } from "@oclif/core";
+
