+++ solana/gatekeeper-lib/docs/classes/GatekeeperNetworkService.html
@@ -115,7 +115,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/83df0519/solana/gatekeeper-lib/src/service/gatekeeper-network-service.ts#L31">service/gatekeeper-network-service.ts:31</a></li>
+									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/f06497c1/solana/gatekeeper-lib/src/service/gatekeeper-network-service.ts#L31">service/gatekeeper-network-service.ts:31</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -159,7 +159,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/83df0519/solana/gatekeeper-lib/src/service/gatekeeper-network-service.ts#L64">service/gatekeeper-network-service.ts:64</a></li>
+									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/f06497c1/solana/gatekeeper-lib/src/service/gatekeeper-network-service.ts#L64">service/gatekeeper-network-service.ts:64</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -197,7 +197,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/83df0519/solana/gatekeeper-lib/src/service/gatekeeper-network-service.ts#L142">service/gatekeeper-network-service.ts:142</a></li>
+									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/f06497c1/solana/gatekeeper-lib/src/service/gatekeeper-network-service.ts#L142">service/gatekeeper-network-service.ts:142</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -241,7 +241,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/83df0519/solana/gatekeeper-lib/src/service/gatekeeper-network-service.ts#L122">service/gatekeeper-network-service.ts:122</a></li>
+									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/f06497c1/solana/gatekeeper-lib/src/service/gatekeeper-network-service.ts#L122">service/gatekeeper-network-service.ts:122</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -273,7 +273,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/83df0519/solana/gatekeeper-lib/src/service/gatekeeper-network-service.ts#L201">service/gatekeeper-network-service.ts:201</a></li>
+									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/f06497c1/solana/gatekeeper-lib/src/service/gatekeeper-network-service.ts#L201">service/gatekeeper-network-service.ts:201</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -305,7 +305,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/83df0519/solana/gatekeeper-lib/src/service/gatekeeper-network-service.ts#L172">service/gatekeeper-network-service.ts:172</a></li>
+									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/f06497c1/solana/gatekeeper-lib/src/service/gatekeeper-network-service.ts#L172">service/gatekeeper-network-service.ts:172</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
