+++ solana/gatekeeper-lib/docs/classes/GatekeeperService.html
@@ -117,7 +117,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/83df0519/solana/gatekeeper-lib/src/service/gatekeeper-service.ts#L34">service/gatekeeper-service.ts:34</a></li>
+									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/f06497c1/solana/gatekeeper-lib/src/service/gatekeeper-service.ts#L34">service/gatekeeper-service.ts:34</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -177,7 +177,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/83df0519/solana/gatekeeper-lib/src/service/gatekeeper-service.ts#L265">service/gatekeeper-service.ts:265</a></li>
+									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/f06497c1/solana/gatekeeper-lib/src/service/gatekeeper-service.ts#L265">service/gatekeeper-service.ts:265</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -215,7 +215,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/83df0519/solana/gatekeeper-lib/src/service/gatekeeper-service.ts#L228">service/gatekeeper-service.ts:228</a></li>
+									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/f06497c1/solana/gatekeeper-lib/src/service/gatekeeper-service.ts#L228">service/gatekeeper-service.ts:228</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -253,7 +253,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/83df0519/solana/gatekeeper-lib/src/service/gatekeeper-service.ts#L159">service/gatekeeper-service.ts:159</a></li>
+									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/f06497c1/solana/gatekeeper-lib/src/service/gatekeeper-service.ts#L159">service/gatekeeper-service.ts:159</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -291,7 +291,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/83df0519/solana/gatekeeper-lib/src/service/gatekeeper-service.ts#L201">service/gatekeeper-service.ts:201</a></li>
+									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/f06497c1/solana/gatekeeper-lib/src/service/gatekeeper-service.ts#L201">service/gatekeeper-service.ts:201</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -329,7 +329,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/83df0519/solana/gatekeeper-lib/src/service/gatekeeper-service.ts#L247">service/gatekeeper-service.ts:247</a></li>
+									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/f06497c1/solana/gatekeeper-lib/src/service/gatekeeper-service.ts#L247">service/gatekeeper-service.ts:247</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -367,7 +367,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/83df0519/solana/gatekeeper-lib/src/service/gatekeeper-service.ts#L280">service/gatekeeper-service.ts:280</a></li>
+									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/f06497c1/solana/gatekeeper-lib/src/service/gatekeeper-service.ts#L280">service/gatekeeper-service.ts:280</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -411,7 +411,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/83df0519/solana/gatekeeper-lib/src/service/gatekeeper-service.ts#L176">service/gatekeeper-service.ts:176</a></li>
+									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/f06497c1/solana/gatekeeper-lib/src/service/gatekeeper-service.ts#L176">service/gatekeeper-service.ts:176</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
