+++ solana/gatekeeper-lib/docs/index.html
@@ -86,7 +86,7 @@
 					<div class="tsd-signature tsd-kind-icon">Charge<wbr>Options<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol">{</span><span class="tsd-signature-symbol">[ </span><span class="tsd-signature-type">k</span><span class="tsd-signature-symbol"> in </span><span class="tsd-signature-type">Action</span><span class="tsd-signature-symbol">]</span><span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">ChargeOption</span><span class="tsd-signature-symbol"> }</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/83df0519/solana/gatekeeper-lib/src/util/types.ts#L17">util/types.ts:17</a></li>
+							<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/f06497c1/solana/gatekeeper-lib/src/util/types.ts#L17">util/types.ts:17</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -101,7 +101,7 @@
 					<div class="tsd-signature tsd-kind-icon">Gatekeeper<wbr>Config<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">TransactionOptions</span><span class="tsd-signature-symbol"> &amp; </span><span class="tsd-signature-symbol">{}</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/83df0519/solana/gatekeeper-lib/src/util/types.ts#L29">util/types.ts:29</a></li>
+							<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/f06497c1/solana/gatekeeper-lib/src/util/types.ts#L29">util/types.ts:29</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
