+++ solana/gatekeeper-lib/package.json
@@ -1,18 +1,11 @@
 {
   "name": "@identity.com/solana-gatekeeper-lib",
-  "description": "Library and CLI to manage Gateway Tokens",
-  "version": "4.1.1-alpha.3",
+  "description": "Library to manage Gateway Tokens",
+  "version": "5.0.0-alpha.1",
   "author": "dankelleher @dankelleher",
-  "bin": {
-    "gateway": "./bin/run"
-  },
   "bugs": "https://github.com/identity-com/on-chain-identity-gateway/issues",
   "dependencies": {
-    "@identity.com/solana-gateway-ts": "0.11.0-alpha.3",
-    "@oclif/core": "^1.5.2",
-    "@oclif/plugin-help": "^5.1.11",
-    "@oclif/plugin-plugins": "^2.0.1",
-    "@oclif/plugin-version": "^1",
+    "@identity.com/solana-gateway-ts": "0.11.0-alpha.4",
     "@solana/buffer-layout": "^4.0.0",
     "@solana/spl-token": "^0.3.5",
     "@solana/web3.js": "^1.61.0",
@@ -22,8 +15,6 @@
     "tslib": "^1"
   },
   "devDependencies": {
-    "@oclif/dev-cli": "^1.26.0",
-    "@oclif/test": "^1.2.8",
     "@types/async-retry": "^1.4.3",
     "@types/bn.js": "^5.1.0",
     "@types/bs58": "^4.0.1",
@@ -40,14 +31,11 @@
     "chai-as-promised": "^7.1.1",
     "chai-subset": "^1.6.0",
     "eslint": "^7.26.0",
-    "eslint-config-oclif": "^3.1.0",
-    "eslint-config-oclif-typescript": "^0.2.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^3.4.0",
     "globby": "^10",
     "mocha": "^5",
     "nyc": "^14",
-    "oclif": "^1.17.0",
     "postinstall-prepare": "^2.0.0",
     "prettier": "^2.3.0",
     "sinon": "^13.0.1",
@@ -91,15 +79,13 @@
   },
   "repository": "git@github.com:identity-com/on-chain-identity-gateway.git",
   "scripts": {
-    "build": "rm -rf dist && tsc -b && oclif-dev manifest && oclif-dev readme && cp src/util/oclif/*.json dist/util/oclif/ && yarn doc",
-    "postpack": "rm -f oclif.manifest.json",
+    "build": "rm -rf dist && tsc -b",
     "posttest": "yarn lint",
     "prepack": "yarn build",
     "start-validator": "solana-test-validator --bpf-program gatem74V238djXdzWnJf94Wo1DcnuGkfijbf3AuBhfs ../program/target/deploy/solana_gateway_program.so --reset",
     "lint": "eslint '{src,test}/**/*.ts' --max-warnings 0",
     "lint:fix": "yarn lint --fix",
     "test": "STAGE=${STAGE:-dev} FORCE_COLOR=1 nyc --extension .ts --reporter=text ts-mocha --timeout 30000 --paths 'test/**/*.ts'",
-    "version": "oclif-dev readme && git add README.md",
     "doc": "yarn typedoc src/index.ts --readme none --excludeNotDocumented",
     "audit-ci": "audit-ci --config audit-ci.json"
   },
