+++ solana/gatekeeper-lib/src/util/index.ts
@@ -1,11 +1,12 @@
 export { airdropTo } from "./account";
 export {
   getConnection,
+  getClusterUrl,
   SentDataTransaction,
   SentTransaction,
   SendableTransaction,
   SendableDataTransaction,
+  ExtendedCluster
 } from "./connection";
 export * as constants from "./constants";
-export * as token from "./token";
 export { ChargeOptions, Action, GatekeeperConfig, TransactionOptions } from "./types";
