+++ solana/gateway-ts/docs/classes/Revoked.html
@@ -111,7 +111,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from Assignable.constructor</p>
 								<ul>
-									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/83df0519/solana/gateway-ts/src/lib/solanaBorsh.ts#L7">lib/solanaBorsh.ts:7</a></li>
+									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/17c33ed6/solana/gateway-ts/src/lib/solanaBorsh.ts#L7">lib/solanaBorsh.ts:7</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -143,7 +143,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from Assignable.encode</p>
 								<ul>
-									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/83df0519/solana/gateway-ts/src/lib/solanaBorsh.ts#L16">lib/solanaBorsh.ts:16</a></li>
+									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/17c33ed6/solana/gateway-ts/src/lib/solanaBorsh.ts#L16">lib/solanaBorsh.ts:16</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">Buffer</span></h4>
@@ -161,7 +161,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from Assignable.decode</p>
 								<ul>
-									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/83df0519/solana/gateway-ts/src/lib/solanaBorsh.ts#L20">lib/solanaBorsh.ts:20</a></li>
+									<li>Defined in <a href="https://github.com/identity-com/on-chain-identity-gateway/blob/17c33ed6/solana/gateway-ts/src/lib/solanaBorsh.ts#L20">lib/solanaBorsh.ts:20</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-type-parameters-title">Type parameters</h4>
